+++ src/proof-of-delivery-view/proof-of-delivery-view.module.js
@@ -30,7 +30,8 @@
         'openlmis-pagination',
         'order',
         'ui.router',
-        'shipment'
+        'shipment',
+        'stock-valid-reason'
     ]);
 
 })();