+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New functionality added in a backwards-compatible manner:
 * [OLMIS-239](https://openlmis.atlassian.net/browse/OLMIS-239): Added print button to shipment page.
 * [OLMIS-1611](https://openlmis.atlassian.net/browse/OLMIS-1611): Added confirm shipment button.
 * [OLMIS-529](https://openlmis.atlassian.net/browse/OLMIS-529): Show lots in recommended sort order on fulfill view.
+* [OLMIS-685](https://openlmis.atlassian.net/browse/OLMIS-685): Added the ability to provide rejection reason for stock.
 
 Improvements:
 * [OLMIS-3876](https://openlmis.atlassian.net/browse/OLMIS-3876): Split navigation and filter forms in manage POD and orders pages